+++ src/node-process/serve.js
@@ -1,21 +1,9 @@
 'use strict';
 
-const Server = require('./Server'),
-    ConnectionDelegate = require('./ConnectionDelegate'),
+const ConsoleInterceptor = require('./NodeInterceptors/ConsoleInterceptor'),
+    Logger = require('./Logger'),
+    Server = require('./Server'),
     DataSerializer = require('./Data/Serializer');
-
-// Instanciate the custom connection delegate
-const connectionDelegate = new (require(process.argv.slice(2)[0]));
-
-// Retrieve the options
-let options = process.argv.slice(2)[1];
-options = options !== undefined ? JSON.parse(options) : {};
-
-// Start the server with the custom connection delegate
-const server = new Server(connectionDelegate, options);
-
-// Write the server port to the process output
-server.started.then(() => server.writePortToOutput());
 
 // Throw unhandled rejections
 process.on('unhandledRejection', error => {
@@ -25,6 +13,28 @@ process.on('unhandledRejection', error => {
 // Output the exceptions in JSON format
 process.on('uncaughtException', error => {
     process.stderr.write(JSON.stringify(DataSerializer.serializeError(error)));
-
     process.exit(1);
 });
+
+// Retrieve the options
+let options = process.argv.slice(2)[1];
+options = options !== undefined ? JSON.parse(options) : {};
+
+// Intercept Node logs
+if (options.log_node_console === true) {
+    ConsoleInterceptor.startInterceptingLogs((type, originalMessage) => {
+        const level = ConsoleInterceptor.getLevelFromType(type);
+        const message = ConsoleInterceptor.formatMessage(originalMessage);
+
+        Logger.log('Node', level, message);
+    });
+}
+
+// Instanciate the custom connection delegate
+const connectionDelegate = new (require(process.argv.slice(2)[0]));
+
+// Start the server with the custom connection delegate
+const server = new Server(connectionDelegate, options);
+
+// Write the server port to the process output
+server.started.then(() => server.writePortToOutput());